+++ pkg/logql/syntax/ast.go
@@ -260,7 +260,7 @@ func (e *PipelineExpr) Pipeline() (log.Pipeline, error) {
 func (e *PipelineExpr) HasFilter() bool {
 	for _, p := range e.MultiStages {
 		switch p.(type) {
-		case *LineFilterExpr, *LabelFilterExpr:
+		case *LineFilterExpr, *LabelFilterExpr, *DistinctFilterExpr:
 			return true
 		default:
 			continue
@@ -631,6 +631,37 @@ func (j *JSONExpressionParser) String() string {
 	return sb.String()
 }
 
+type DistinctFilterExpr struct {
+	labels []string
+	implicit
+}
+
+func newDistinctFilterExpr(labels []string) *DistinctFilterExpr {
+	return &DistinctFilterExpr{
+		labels: labels,
+	}
+}
+
+func (e *DistinctFilterExpr) Shardable() bool { return false }
+
+func (e *DistinctFilterExpr) Walk(f WalkFn) { f(e) }
+
+func (e *DistinctFilterExpr) Stage() (log.Stage, error) {
+	return log.NewDistinctFilter(e.labels)
+}
+
+func (e *DistinctFilterExpr) String() string {
+	var sb strings.Builder
+	sb.WriteString(fmt.Sprintf("%s %s ", OpPipe, OpFilterDistinct))
+	for i, label := range e.labels {
+		sb.WriteString(label)
+		if i+1 != len(e.labels) {
+			sb.WriteString(",")
+		}
+	}
+	return sb.String()
+}
+
 type internedStringSet map[string]struct {
 	s  string
 	ok bool
@@ -692,7 +723,7 @@ func simplifyRegexMatcher(typ labels.MatchType, name, value string) *labels.Matc
 	}
 	reg = reg.Simplify()
 
-	m, ok := simplify(typ, name, value, reg)
+	m, ok := simplify(typ, name, reg)
 	if !ok {
 		util.AllNonGreedy(reg)
 		return labels.MustNewMatcher(typ, name, reg.String())
@@ -702,7 +733,7 @@ func simplifyRegexMatcher(typ labels.MatchType, name, value string) *labels.Matc
 }
 
 // simplify will return an equals matcher if there is a regex matching a literal
-func simplify(typ labels.MatchType, name, value string, reg *syntax.Regexp) (*labels.Matcher, bool) {
+func simplify(typ labels.MatchType, name string, reg *syntax.Regexp) (*labels.Matcher, bool) {
 	switch reg.Op {
 	case syntax.OpLiteral:
 		if !util.IsCaseInsensitive(reg) {
@@ -710,7 +741,7 @@ func simplify(typ labels.MatchType, name, value string, reg *syntax.Regexp) (*la
 			if typ == labels.MatchNotRegexp {
 				t = labels.MatchNotEqual
 			}
-			return labels.MustNewMatcher(t, name, value), true
+			return labels.MustNewMatcher(t, name, string(reg.Rune)), true
 		}
 		return nil, false
 	}
@@ -903,6 +934,8 @@ const (
 
 	// function filters
 	OpFilterIP = "ip"
+
+	OpFilterDistinct = "distinct"
 
 	// drop labels
 	OpDrop = "drop"