+++ production/helm/loki/CHANGELOG.md
@@ -13,6 +13,17 @@ Entries should include a reference to the pull request that introduced the chang
 
 [//]: # (<AUTOMATED_UPDATES_LOCATOR> : do not remove this line. This locator is used by the CI pipeline to automatically create a changelog entry for each new Loki release. Add other chart versions and respective changelog entries bellow this line.)
 
+## 5.5.0
+
+- [CHANGE] Changed version of Grafana Enterprise Logs to v1.7.2
+
+
+## 5.4.0
+
+- [CHANGE] Changed version of Loki to 2.8.2
+
+
+- [CHANGE] Change default GEL and Loki versions to 1.7.1 and 2.8.1 respectively
 - [BUGFIX] Fix dns port in network-policy
 
 ## 4.10.0